feat(Chat): token union type — badge config + custom render() escape hatch
Token type is now a union:
- Badge config (recommended): { value, label, variant?, icon? }
  Structured, autocomplete-friendly, renders XDSBadge via portal.
- Custom render (escape hatch): { value, render: () => ReactNode }
  Full control — tooltips, hovercards, clickable tokens, anything.

isCustomToken() type guard routes to the right path in both the
portal rendering and XDSChatComposerTokenElement.

Stories:
- Renamed file to ChatComposerInput.stories.tsx
- Added 6 basic input stories (standalone, placeholder, disabled,
  max rows, message history, file drop)
- CustomRenderItem now shows icon on badge config
- New CustomRender story: clickable tokens with avatar icons
- TokenVariants: different badge colors per trigger
- 13 total stories

Also fixed Chat.doc.mjs — cssVars moved to theming.vars (correct
ComponentDoc type).
